Fix Rollbar# 19464 - an exception is thrown when cAPI price data is not expressed as integers (https://rollbar.com/EDCD/EDDI/items/19464/) ```json {    "id": 128666747,    "name": "LavianBrandy",    "legality": "I",    "buyPrice": 745.8000000000001,    "sellPrice": 12833.203434782608,    "meanPrice": 10365,    "demandBracket": 0,    "stockBracket": 0,    "stock": 0,    "demand": 0,    "statusFlags": [],    "categoryname": "Narcotics",    "locName": "Lavian Brandy"   } ```

+++ DataDefinitions/CommodityDefinition.cs
@@ -448,7 +448,7 @@ namespace EddiDataDefinitions
 
         // The average price of a commodity can change - thus this cannot be read only.
         // Instead, this value should be updated whenever revised data is received.
-        public int avgprice { get; set; }
+        public decimal avgprice { get; set; }
 
         // dummy used to ensure that the static constructor has run
         public CommodityDefinition() : this(0, null, "", Unknown)
